modules/ProvisioningModule/requirements.txt
azure-iot-device
toml
black; python_version >= '3.6'
flake8
colorama